+++ libnopegl/src/ngpu/opengl/glcontext_egl.c
@@ -495,7 +495,7 @@ try_again:;
                 LOG(ERROR, "no Wayland display specified");
                 return NGL_ERROR_INVALID_ARG;
             }
-            egl->wl_egl_window = wl_egl_window_create(wl_surface, ctx->width, ctx->height);
+            egl->wl_egl_window = wl_egl_window_create(wl_surface, (EGLint)ctx->width, (EGLint)ctx->height);
             if (!egl->wl_egl_window) {
                 LOG(ERROR, "could not create Wayland EGL window");
                 return NGL_ERROR_EXTERNAL;
@@ -576,7 +576,7 @@ static void egl_uninit(struct glcontext *ctx)
 #endif
 }
 
-static int egl_resize(struct glcontext *ctx, int32_t width, int32_t height)
+static int egl_resize(struct glcontext *ctx, uint32_t width, uint32_t height)
 {
     struct egl_priv *egl = ctx->priv_data;
 
@@ -605,8 +605,8 @@ static int egl_resize(struct glcontext *ctx, int32_t width, int32_t height)
         LOG(ERROR, "could not query surface dimensions: %s", ngli_eglGetErrorStr());
         return NGL_ERROR_EXTERNAL;
     }
-    ctx->width = cur_width;
-    ctx->height = cur_height;
+    ctx->width = (uint32_t)cur_width;
+    ctx->height = (uint32_t)cur_height;
 
     return 0;
 }